// src/prisma/generated/prisma/models/ChannelRole.ts
export type ChannelRoleWhereInput = {
  AND?: Prisma.ChannelRoleWhereInput | Prisma.ChannelRoleWhereInput[]
  OR?: Prisma.ChannelRoleWhereInput[]
  NOT?: Prisma.ChannelRoleWhereInput | Prisma.ChannelRoleWhereInput[]
  id?: Prisma.IntFilter<"ChannelRole"> | number
  channelId?: Prisma.IntFilter<"ChannelRole"> | number
  roleId?: Prisma.IntFilter<"ChannelRole"> | number
  permissions?: Prisma.JsonFilter<"ChannelRole">
  channel?: Prisma.XOR<Prisma.ChannelScalarRelationFilter, Prisma.ChannelWhereInput>
  role?: Prisma.XOR<Prisma.RoleScalarRelationFilter, Prisma.RoleWhereInput>
}

export type ChannelRoleOrderByWithRelationInput = {
  id?: Prisma.SortOrder
  channelId?: Prisma.SortOrder
  roleId?: Prisma.SortOrder
  permissions?: Prisma.SortOrder
  channel?: Prisma.ChannelOrderByWithRelationInput
  role?: Prisma.RoleOrderByWithRelationInput
}

export type ChannelRoleWhereUniqueInput = Prisma.AtLeast<{
  id?: number
  AND?: Prisma.ChannelRoleWhereInput | Prisma.ChannelRoleWhereInput[]
  OR?: Prisma.ChannelRoleWhereInput[]
  NOT?: Prisma.ChannelRoleWhereInput | Prisma.ChannelRoleWhereInput[]
  channelId?: Prisma.IntFilter<"ChannelRole"> | number
  roleId?: Prisma.IntFilter<"ChannelRole"> | number
  permissions?: Prisma.JsonFilter<"ChannelRole">
  channel?: Prisma.XOR<Prisma.ChannelScalarRelationFilter, Prisma.ChannelWhereInput>
  role?: Prisma.XOR<Prisma.RoleScalarRelationFilter, Prisma.RoleWhereInput>
}, "id">
// ...
export type ChannelRoleSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  channelId?: boolean
  roleId?: boolean
  permissions?: boolean
  channel?: boolean | Prisma.ChannelDefaultArgs<ExtArgs>
  role?: boolean | Prisma.RoleDefaultArgs<ExtArgs>
}, ExtArgs["result"]["channelRole"]>
// ...
export type ChannelRoleOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "channelId" | "roleId" | "permissions", ExtArgs["result"]["channelRole"]>
export type ChannelRoleInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  channel?: boolean | Prisma.ChannelDefaultArgs<ExtArgs>
  role?: boolean | Prisma.RoleDefaultArgs<ExtArgs>
}
// ...
export type ChannelRoleFindUniqueArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the ChannelRole
   */
  select?: Prisma.ChannelRoleSelect<ExtArgs> | null
  /**
   * Omit specific fields from the ChannelRole
   */
  omit?: Prisma.ChannelRoleOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.ChannelRoleInclude<ExtArgs> | null
  /**
   * Filter, which ChannelRole to fetch.
   */
  where: Prisma.ChannelRoleWhereUniqueInput
}